feat: update email settings
